API documentation

Build tables for Kanones dataset:

LSJMining.adjs1_2 — Function

Extract 1st-2nd declension adjectives from v and write a stems table to the Kanones dataset in target.

adjs1_2(v, target)

LSJMining.verbs — Function

Extract all verbs from LSJ and format delimited-text representation for Kanones.

verbs(v, registry, target)

LSJMining.splitmorphemes — Function

Split the verb string s into morpheme units.

splitmorphemes(s, breathsdict; withfailure)

Members

breathsdict: mapping of breathless forms to fully accented forms

Filters to select regular verb classes:

LSJMining.vowelverb — Function

True if label for m is a vowel-stem verb pattern.

LSJMining.liquidverb — Function

True if label for m is a liquid verb pattern.

LSJMining.stopverb — Function

True if label for m is a stop verb pattern.

LSJMining.contractverb — Function

True if label for m is a contract verb pattern.

LSJMining.izwverb — Function

True if label for m is a regular verb pattern in -ίζω.

LSJMining.sigmaverb — Function

True if label for m is a stop verb pattern.

LSJMining.numiverb — Function

True if label for m is a regular verb pattern in -νυμι.

LSJMining.irregmiverb — Function

True if label for m is an iregular -μι verb pattern.

LSJMining.irregomega — Function

True if label for m is an iregular -ω verb pattern.

Utilities for managing files:

LSJMining.loadmorphdata — Function

Load morphology data from files in cex directory of repository, and filter for entries with valid orthography only.

loadmorphdata(cexdir, kroot)

Load all morphology data from files in cex directory of repository.

loadmorphdata(cexdir)